Refactor common filters to common folder
