+++ components/dashboard/sections/contributions/filters.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
-import { defineMessage, FormattedMessage } from 'react-intl';
+import { defineMessage } from 'react-intl';
 import { z } from 'zod';
 
-import type { FilterComponentConfigs, FilterConfig, FiltersToVariables } from '../../../../lib/filters/filter-types';
+import type { FilterComponentConfigs, FiltersToVariables } from '../../../../lib/filters/filter-types';
 import { integer, isMulti } from '../../../../lib/filters/schemas';
 import {
   ContributionFrequency,
   Currency,
   DashboardRecurringContributionsQueryVariables,
-  ExpectedFundsFilter,
   HostContributionsQueryVariables,
   OrderStatus,
 } from '../../../../lib/graphql/types/v2/graphql';
@@ -17,9 +16,8 @@ import { sortSelectOptions } from '../../../../lib/utils';
 
 import { amountFilter } from '../../filters/AmountFilter';
 import ComboSelectFilter from '../../filters/ComboSelectFilter';
-import { DateFilter, dateFilter } from '../../filters/DateFilter';
-import { DateFilterValue } from '../../filters/DateFilter/DateFilterValue';
-import { dateFilterSchema, dateToVariables } from '../../filters/DateFilter/schema';
+import { dateFilter, expectedDateFilter } from '../../filters/DateFilter';
+import { expectedFundsFilter } from '../../filters/ExpectedFundsFilter';
 import { orderByFilter } from '../../filters/OrderFilter';
 import { searchFilter } from '../../filters/SearchFilter';
 
@@ -38,16 +36,6 @@ const i18nOrderType = (intl, value) => {
 
 const PAGE_SIZE = 20;
 
-const expectedDateFilter: FilterConfig<z.infer<typeof dateFilterSchema>> = {
-  schema: dateFilterSchema,
-  toVariables: value => dateToVariables(value, 'expectedDate'),
-  filter: {
-    labelMsg: defineMessage({ defaultMessage: 'Expected Date', id: 'vNC2dX' }),
-    Component: DateFilter,
-    valueRenderer: DateFilterValue,
-  },
-};
-
 export const schema = z.object({
   limit: integer.default(PAGE_SIZE),
   offset: integer.default(0),
@@ -55,7 +43,7 @@ export const schema = z.object({
   searchTerm: searchFilter.schema,
   date: dateFilter.schema,
   expectedDate: expectedDateFilter.schema,
-  expectedFundsFilter: z.nativeEnum(ExpectedFundsFilter).optional(),
+  expectedFundsFilter: expectedFundsFilter.schema,
   amount: amountFilter.schema,
   status: isMulti(z.nativeEnum(OrderStatus)).optional(),
   type: z.nativeEnum(OrderTypeFilter).optional(),
@@ -102,22 +90,7 @@ export const filters: FilterComponentConfigs<FilterValues, FilterMeta> = {
   searchTerm: searchFilter.filter,
   date: dateFilter.filter,
   expectedDate: expectedDateFilter.filter,
-  expectedFundsFilter: {
-    labelMsg: defineMessage({ defaultMessage: 'Expected Funds', id: 'ExpectedFunds' }),
-    Component: ({ valueRenderer, intl, ...props }) => {
-      const options = React.useMemo(
-        () => Object.values(ExpectedFundsFilter).map(value => ({ label: valueRenderer({ intl, value }), value })),
-        [intl, valueRenderer],
-      );
-      return <ComboSelectFilter isMulti={false} options={options} {...props} />;
-    },
-    valueRenderer: ({ value }) =>
-      value === ExpectedFundsFilter.MANUAL ? (
-        <FormattedMessage defaultMessage="Created by contributors" id="wa5frV" />
-      ) : (
-        <FormattedMessage defaultMessage="Created by host admin" id="See4A8" />
-      ),
-  },
+  expectedFundsFilter: expectedFundsFilter.filter,
   amount: { ...amountFilter.filter, labelMsg: defineMessage({ id: 'TotalAmount', defaultMessage: 'Total amount' }) },
   orderBy: orderByFilter.filter,
   status: {
